requirements.txt
pyyaml
Jinja2
paramiko
'git+https://github.com/facebookresearch/detectron2.git'